Move webhook package
- `rpc` -> `server`
